build: Setup linting
This commit adds a basic eslint configuration, and sets up a gulp task
to run it.

To use it, install the gulp CLI globally with `npm install -g gulp-cli`.
Then, in the root directory of this repo, run `gulp lint`. A target for
automatically fixing errors is also provided as `gulp lint-fix`.