spark: Use public IP for application UI
This commit adds functionality to fix the hostname used to link
to the Spark driver API (from the Spark Master API) to use the
master's public IP address.  Spark allows users to configure the
hostname by setting an environment variable, but this is tricky to
do with Kelda because we don't know the correct value of
this until after the container boots (so we can't set the
environment variable in the blueprint).  If a user used the existing
blueprint, for the UI links to be correct, the user would need to
set the SPARK_PUBLIC_DNS environment variable herself before running
a job; e.g., by running:

$ SPARK_PUBLIC_DNS=1.2.3.4 ./spark-shell

when ssh'ed into the Spark container, where 1.2.3.4 would need to be
replaced with the public IP of the container running the master.

This commit fixes this problem by adding a spark-env.sh file that sets the
environment variable.  spark-env.sh is a useful (and expected) way to set
environment variables for Spark jobs, because the Spark job running scripts
source environment variables in this file when running a job.