spark: Allow all nodes to access public on 80
Even if a spark node doesn't expose its UI to public, the container
still checks it's public IP by accessing an external service on 80.
The best solution to this problem is to tell the spark containers
their IP through an environment variable, but until that's possible
this patch just opens up the port.