spark: Enable the history server
This commit updates Spark to start the history server. The history
server runs on port 18081 on the master, and displays the application
UI (that's usually available on port 4040) for applications
that have completed.  It's useful for debugging jobs that ran in
the past.

Currently, we use the default configuration, which saves data on the
local filesystem on the master. In the future, we may want to store
the history server data in HDFS, in which case we can run the history
server in its own container.