blueprint: Add log4j.properties in the blueprint
This commit moves the log4j.properties file so that it is no
longer copied by the Dockerfile, and is instead copied by the
blueprint.  This is useful so that a user can modify log4j.properties
at runtime to change the logging level, and also makes it so that
all of the configuration files are copied in a consistent way.