blueprint: Run driver in its own container
The driver is its own process, so this commit runs the driver
as a separate container. Ideally the driver and the master should
be on the same machine.  That appears to happen by default with
the current scheduler, but once Kelda placement policies support
it, the blueprint should explicitly specify that these should be
on the same machine.

This commit also updates the readme to describe how to interact with
the driver.