blueprint: Add function to compute memory
It's a little tricky to compute the ideal amount of memory that
should be allocated to each Spark worker. This commit adds a function
to the spark module that accepts a machine, and returns the amount
of memory Spark should be given for that machine.

As part of this change, this commit also changes the blueprint to
explicitly tell each worker how much memory it has available.
Spark has two components that each have a memory configuration: there's
a Worker, which runs on each worker machine, and an Executor.  The Worker
manages everything running on the machine; each application has
its own Executors that get launched by the Worker.  Previously, Kelda
computed how much memory to allocate to each Executor, but the Worker
calculated its memory on its own (by trying to read it from the file-
system). This could lead to problems where Kelda configured Executors to
have more memory than Workers thought was available, in which case the
application would hang because Spark thought none of the workers had
enough memory to start an Executor. To fix this problem, this commit
updates the blueprint to explicitly configure the memory on both the
Executor and the Worker, so that they're consistent.