blueprint: Set SPARK_PUBLIC_DNS using kelda.hostIP
The new kelda.hostIP functionality allows blueprints to set an
environment variable based on the public IP address of the machine
where the container is launched. This commit uses that functionality
to remove spark-env.sh (and the need to expose port 80 so that
Spark containers can learn their own address).