blueprint: Fix memory configuration
This commit fixes a few small problems related to the memory configuration:
1) The blueprint treated machine RAM size as GB, which has now been
corrected to GiB.
2) The memory calculation could previously result in a negative value
which is both nonsensical and would cause an error when executing the
Spark worker command, since the minus would cause the memory size to be
treated as a flag rather than an argument.
3) We previously used "MB" as our memory unit in the worker command line
command, but Spark encourages using "M" (which is also less misleading,
given that the real unit is MiB).